+++ domain/src/main/java/com/mordansoft/homebank/domain/repo/ProfitRepo.kt
@@ -5,6 +5,6 @@ import com.mordansoft.homebank.domain.model.Profit
 interface ProfitRepo {
     suspend fun updateProfit(profit: Profit)
     suspend fun getProfitById(profitId: Long): Profit
-    suspend fun getMainProfits(periodId : Long): ArrayList<Profit>
+    suspend fun getMainProfits(periodId : Int): ArrayList<Profit>
     suspend fun insertTestProfit(){    }
 }
